WIP: LKI in attack flow for whistling birds
