+++ server/game/core/gameSystem/CardTargetSystem.ts
@@ -5,6 +5,7 @@ import { EffectName, EventName, GameStateChangeRequired, WildcardCardType, ZoneN
 import type { IGameSystemProperties as IGameSystemProperties } from './GameSystem';
 import { GameSystem as GameSystem } from './GameSystem';
 import { GameEvent } from '../event/GameEvent';
+import { addLastKnownInformationToEvent, buildLastKnownInformation } from '../event/LastKnownInformation';
 import { EnumHelpers } from '../utils/EnumHelpers';
 import { Helpers } from '../utils/Helpers';
 import { Contract } from '../utils/Contract';
@@ -152,7 +153,7 @@ export abstract class CardTargetSystem<TContext extends AbilityContext = Ability
 
         const event = this.createEvent(nonArrayTarget, context, additionalProperties);
         if (addLastKnownInformation) {
-            (event as any).lastKnownInformation = this.buildLastKnownInformation(nonArrayTarget);
+            (event as any).lastKnownInformation = buildLastKnownInformation(nonArrayTarget);
         }
         this.updateEvent(event, nonArrayTarget, context, additionalProperties);
         return event;
@@ -200,7 +201,7 @@ export abstract class CardTargetSystem<TContext extends AbilityContext = Ability
                 card
             });
 
-            this.addLastKnownInformationToEvent(onCardLeavesPlayEvent, card);
+            addLastKnownInformationToEvent(onCardLeavesPlayEvent, card);
 
             let contingentEvents = [onCardLeavesPlayEvent];
